@@ -20,7 +20,7 @@
     ],
     "homepage": "https://docs.sonata-project.org/projects/SonataBlockBundle",
     "require": {
-        "php": "^7.4 || ^8.0",
+        "php": "^8.0",
         "doctrine/collections": "^1.6 || ^2.0",
         "doctrine/common": "^2.7.1 || ^3.0",
         "psr/cache": "^1.0 || ^2.0 || ^3.0",
@@ -28,19 +28,19 @@
         "psr/log": "^1.1 || ^2.0 || ^3.0",
         "sonata-project/cache": "^2.0",
         "sonata-project/form-extensions": "^1.4",
-        "symfony/asset": "^4.4 || ^5.4 || ^6.0",
-        "symfony/config": "^4.4 || ^5.4 || ^6.0",
-        "symfony/console": "^4.4 || ^5.4 || ^6.0",
-        "symfony/dependency-injection": "^4.4 || ^5.4 || ^6.0",
-        "symfony/event-dispatcher": "^4.4 || ^5.4 || ^6.0",
+        "symfony/asset": "^4.4 || ^5.4 || ^6.2",
+        "symfony/config": "^4.4 || ^5.4 || ^6.2",
+        "symfony/console": "^4.4 || ^5.4 || ^6.2",
+        "symfony/dependency-injection": "^4.4 || ^5.4 || ^6.2",
+        "symfony/event-dispatcher": "^4.4 || ^5.4 || ^6.2",
         "symfony/event-dispatcher-contracts": "^1.1 || ^2.0 || ^3.0",
-        "symfony/form": "^4.4 || ^5.4 || ^6.0",
-        "symfony/framework-bundle": "^4.4 || ^5.4 || ^6.0",
-        "symfony/http-foundation": "^4.4 || ^5.4 || ^6.0",
-        "symfony/http-kernel": "^4.4 || ^5.4 || ^6.0",
-        "symfony/options-resolver": "^4.4 || ^5.4 || ^6.0",
-        "symfony/twig-bundle": "^4.4 || ^5.4 || ^6.0",
-        "symfony/validator": "^4.4 || ^5.4 || ^6.0",
+        "symfony/form": "^4.4 || ^5.4 || ^6.2",
+        "symfony/framework-bundle": "^4.4 || ^5.4 || ^6.2",
+        "symfony/http-foundation": "^4.4 || ^5.4 || ^6.2",
+        "symfony/http-kernel": "^4.4 || ^5.4 || ^6.2",
+        "symfony/options-resolver": "^4.4 || ^5.4 || ^6.2",
+        "symfony/twig-bundle": "^4.4 || ^5.4 || ^6.2",
+        "symfony/validator": "^4.4 || ^5.4 || ^6.2",
         "twig/twig": "^2.12.1 || ^3.0"
     },
     "require-dev": {
@@ -59,10 +59,12 @@
         "psalm/plugin-symfony": "^4.0",
         "rector/rector": "^0.15",
         "sonata-project/doctrine-extensions": "^1.10.1",
-        "symfony/browser-kit": "^4.4 || ^5.4 || ^6.0",
-        "symfony/debug": "^4.4 || ^5.4 || ^6.0",
-        "symfony/phpunit-bridge": "^6.1",
-        "symfony/stopwatch": "^4.4 || ^5.4 || ^6.0",
+        "symfony/browser-kit": "^4.4 || ^5.4 || ^6.2",
+        "symfony/debug": "^4.4 || ^5.4 || ^6.2",
+        "symfony/phpunit-bridge": "^6.2",
+        "symfony/routing": "^4.4.12 || ^5.4 || ^6.2",
+        "symfony/stopwatch": "^4.4 || ^5.4 || ^6.2",
+        "symfony/yaml": "^4.4 || ^5.4 || ^6.2",
         "vimeo/psalm": "^4.1"
     },
     "conflict": {
