@@ -26,12 +26,6 @@ use Symfony\Component\HttpFoundation\Response;
  */
 final class BlockRenderer implements BlockRendererInterface
 {
-    private BlockServiceManagerInterface $blockServiceManager;
-
-    private StrategyManagerInterface $exceptionStrategyManager;
-
-    private ?LoggerInterface $logger;
-
     /**
      * This property hold the last response available from the child or sibling block
      * The cacheable attributes must be cascaded to the parent.
@@ -44,13 +38,10 @@ final class BlockRenderer implements BlockRendererInterface
      * @param LoggerInterface              $logger                   Logger class
      */
     public function __construct(
-        BlockServiceManagerInterface $blockServiceManager,
-        StrategyManagerInterface $exceptionStrategyManager,
-        ?LoggerInterface $logger = null
+        private BlockServiceManagerInterface $blockServiceManager,
+        private StrategyManagerInterface $exceptionStrategyManager,
+        private ?LoggerInterface $logger = null
     ) {
-        $this->blockServiceManager = $blockServiceManager;
-        $this->exceptionStrategyManager = $exceptionStrategyManager;
-        $this->logger = $logger;
     }
 
     public function render(BlockContextInterface $blockContext, ?Response $response = null): Response
